refactor: network-config ocap module
