+++ packages/agoric-cli/src/commands/psm.js
@@ -4,7 +4,10 @@
 import { asPercent, makeRpcUtils, storageHelper } from '@agoric/client-utils';
 import { Offers } from '@agoric/inter-protocol/src/clientSupport.js';
 import { Command } from 'commander';
+import { getNetworkConfig } from '../lib/network-config.js';
 import { outputExecuteOfferAction } from '../lib/wallet.js';
+
+const networkConfig = await getNetworkConfig({ env: process.env, fetch });
 
 // Adapted from https://gist.github.com/dckc/8b5b2f16395cb4d7f2ff340e0bc6b610#file-psm-tool
 
@@ -59,7 +62,7 @@ export const makePsmCommand = logger => {
   );
 
   const rpcTools = async () => {
-    const utils = await makeRpcUtils({ fetch });
+    const utils = await makeRpcUtils({ fetch }, networkConfig);
 
     const lookupPsmInstance = ([minted, anchor]) => {
       const name = `psm-${minted}-${anchor}`;
