+++ packages/agoric-cli/src/commands/wallet.js
@@ -11,7 +11,6 @@ import {
 import {
   fmtRecordOfLines,
   makeRpcUtils,
-  networkConfig,
   summarize,
 } from '@agoric/client-utils';
 import { execFileSync } from 'child_process';
@@ -24,7 +23,10 @@ import {
   fetchSwingsetParams,
   normalizeAddressWithOptions,
 } from '../lib/chain.js';
+import { getNetworkConfig } from '../lib/network-config.js';
 import { coalesceWalletState, getCurrent } from '../lib/wallet.js';
+
+const networkConfig = await getNetworkConfig({ env: process.env, fetch });
 
 const SLEEP_SECONDS = 3;
 
@@ -106,7 +108,7 @@ export const makeWalletCommand = async command => {
     .action(async function (opts) {
       const offerStr = fs.readFileSync(opts.file).toString();
 
-      const { unserializer } = await makeRpcUtils({ fetch });
+      const { unserializer } = await makeRpcUtils({ fetch }, networkConfig);
 
       const offerObj = unserializer.fromCapData(JSON.parse(offerStr));
       console.log(offerObj);
@@ -121,7 +123,7 @@ export const makeWalletCommand = async command => {
     .action(async function (opts) {
       const offerStr = fs.readFileSync(opts.offer).toString();
 
-      const { unserializer } = await makeRpcUtils({ fetch });
+      const { unserializer } = await makeRpcUtils({ fetch }, networkConfig);
 
       const offerObj = unserializer.fromCapData(JSON.parse(offerStr));
       console.log(offerObj.offer.id);
@@ -159,7 +161,7 @@ export const makeWalletCommand = async command => {
     .command('list')
     .description('list all wallets in vstorage')
     .action(async function () {
-      const { vstorage } = await makeRpcUtils({ fetch });
+      const { vstorage } = await makeRpcUtils({ fetch }, networkConfig);
       const wallets = await vstorage.keys('published.wallet');
       process.stdout.write(wallets.join('\n'));
     });
@@ -173,16 +175,18 @@ export const makeWalletCommand = async command => {
       normalizeAddress,
     )
     .action(async function (opts) {
-      const { agoricNames, unserializer, readLatestHead } = await makeRpcUtils({
-        fetch,
-      });
+      const { agoricNames, unserializer, readLatestHead } = await makeRpcUtils(
+        {
+          fetch,
+        },
+        networkConfig,
+      );
 
       const leader = makeLeader(networkConfig.rpcAddrs[0]);
       const follower = await makeFollower(
         `:published.wallet.${opts.from}`,
         leader,
         {
-          // @ts-expect-error xxx
           unserializer,
         },
       );
